Defaults

function ishappy(me::Person, circumstance!::Function, happy::Bool=true, args...; kwargs...)
    circumstance!(me, args...; kwargs...)
    me.happyIndexChange > 0 ? return happy : return !happy
end

Dictionaries

Dictionary, n. A malevolent literacy device for cramping the growth of a language and making it hard and inelastic.

Ambrose Bierce on the evils of dictionaries

We've define a set of default dictionaries to make life easier and standardise the extra information users may want to add to their exercises and set schemes. These dictionaries can be expanded by the users using the push!() and insert!() intrinsics. Pull requests adding more entries to the defaul dictionaries are always welcome.

Note

None of these dictionaries are explicitly exported by Lifting.jl. Use them with the Lifting. prefix or import them explicitly.

Lifting.Lifting_Aux — Constant

Lifting_Aux::Dict{String, Dict{String, String}}

Is a dictionary of dictionaries containing non-essential information about exercises and progressions. Its default keys are,

julia> keys(Lifting_Aux)
Base.KeySet for a Dict{String,Dict{String,String}} with 5 entries. Keys:
  "modality"
  "setType"
  "muscles"
  "equipmentSize"
  "equipment"

where each entry is a dictionary with its own keys and values.

Lifting.Lifting_Exercise_Names — Constant

Lifting_Exercise_Names::Dict{String, Dict{String, String}}

Is a dictionary of dictionaries containing lifting exercises. Its default keys are,

julia> import: Lifting_Exercise_Names
julia> keys(Lifting_Exercise_Names)
Base.KeySet for a Dict{String,Dict{String,String}} with 10 entries. Keys:
  "grip"
  "bench"
  "squat"
  "press"
  "isometric"
  "core"
  "pull"
  "row"
  "deadlift"
  "lowBack"

where each entry is a dictionary with its own keys and values.

Lifting.Lifting_Progressions — Constant

Lifting_Progressions::Dict{String, Progression}

Dictionary with preloaded progressions. These are just the progressions I use in my current programme.

julia> import Lifting: Lifting_Progressions
julia> keys(Lifting_Progressions)
Base.KeySet for a Dict{String,Progression} with 14 entries. Keys:
  "AMRAP_Cali"
  "nSuns_6Day_OHP"
  "AMRAP_Grip"
  "l23_lmh"
  "m23_lmh"
  "h23_lmh"
  "CAP3_Bench_T1"
  "CAP3_Row_T1"
  "CAP3_Squat_T2"
  "CAP3_Deadlift_T1"
  "CAP3_Squat_T1"
  "CAP3_Bench_T2"
  "CAP3_Deadlift_T2"
  "CAP3_Row_T2"

Lifting.Lifting_Programmes — Constant

Lifting_Programmes::Dict{String, Programme}

Dictionary with preloaded programmes. These are the programmes I've created and added to the dictionary.

julia> import Lifting: Lifting_Programmes
julia> keys(Lifting_Programmes)
Base.KeySet for a Dict{String,Programme} with 1 entry. Keys:
  "nSunsCAP3_OHP_6Day_LP"