Functions and Methods

Utility

There are various utility functions and methods that enable the programme to work, these can be treated as black boxes.

Base.push! — Function

push!(
    A::AbstractArray{T, 1} where {T},
    exercise::Exercise,
    progression::Progression,
    i::Integer = 1,
)

Calculates set weights using the exercise.trainingMax and progression.setScheme. It also unrolls the relevant data into a named tuple for each day.